[IMP] urban: submodules urbvan and mexico updated Urbvan: Food Voucher rules added. In urbvan they calculate the food voucher a little different than the localization, so here we are overriding that rules just for them.
mexico: Sometimes a public holiday (Feriado) is not paid in its corresponding
payslip period, for example, if the day off is on the 15th of the
month and it is Monday, normally the companies pays the Friday before
so the payslip is validated before the attendance for that day
so the days is never paid.

Adding an manual input to the public holiday salary rule we cover
this case.

Task#43388